code refactoring with oop concepts
